// solution/inventory-app/src/generated/models/Inv2_inventoryitemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Inv2_inventoryitemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Inv2_inventoryitemsstatecode = keyof typeof Inv2_inventoryitemsstatecode;
export const Inv2_inventoryitemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Inv2_inventoryitemsstatuscode = keyof typeof Inv2_inventoryitemsstatuscode;

export interface Inv2_inventoryitemsBase {
  importsequencenumber?: number;
  "inv2_CategoryId@odata.bind"?: string;
  inv2_costperitem?: number;
  inv2_image?: string;
  inv2_inventoryitemid: string;
  inv2_itemname?: string;
  inv2_onhandcount?: number;
  inv2_reorderthreshold?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Inv2_inventoryitemsstatecode;
  statuscode?: Inv2_inventoryitemsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Inv2_inventoryitems extends Inv2_inventoryitemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  inv2_categoryidname?: string;
  inv2_costperitem_base?: number;
  inv2_image_timestamp?: number;
  inv2_image_url?: string;
  inv2_imageid?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  inv2_categoryid?: object;
  _inv2_categoryid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}
export type Inv2_inventoryitemsUploadColumnName = 'inv2_image';