// solution/inventory-app/src/generated/models/Inv2_itemcategoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Inv2_itemcategoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Inv2_itemcategoriesstatecode = keyof typeof Inv2_itemcategoriesstatecode;
export const Inv2_itemcategoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Inv2_itemcategoriesstatuscode = keyof typeof Inv2_itemcategoriesstatuscode;

export interface Inv2_itemcategoriesBase {
  importsequencenumber?: number;
  inv2_categoryname?: string;
  inv2_itemcategoryid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Inv2_itemcategoriesstatecode;
  statuscode?: Inv2_itemcategoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Inv2_itemcategories extends Inv2_itemcategoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}